Move get_valid_target_directories and comment on problems reusing it
